+++ client/ui/Credits.jsx
@@ -33,8 +33,10 @@ export default class Credits extends Component {
 
                       <ul>
                         <b>6.4 </b>
+                        <li>Multi-layer global filter search added</li>
                         <li>Aggregate section updated to allow layer and sublayer filtering in the query</li>
                         <li>Suggest support for query layer and sublayer filtering added</li>
+                        <li>Help suggestions updated to showcase the power of layer and sublayer filtering</li>
                       </ul>
 
                       <ul>