+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 6.4:
   * Aggregate section updated to allow layer and sublayer filtering in the query
   * Suggest support for query layer and sublayer filtering added
+  * Help suggestions updated to showcase the power of layer and sublayer filtering
+  * Multi-layer global filter search added
 
 ## 6.34:
   * Packages updated